Move E2EE modules
