--- apps/meteor/app/e2e/client/events.ts
@@ -1,7 +1,0 @@
-import { Accounts } from 'meteor/accounts-base';
-
-import { e2e } from './rocketchat.e2e';
-
-Accounts.onLogout(() => {
-	void e2e.stopClient();
-});
