+++ apps/meteor/client/lib/chats/flows/uploadFiles.ts
@@ -2,11 +2,11 @@ import type { IMessage, FileAttachmentProps, IE2EEMessage, IUpload } from '@rock
 import { isRoomFederated } from '@rocket.chat/core-typings';
 import { imperativeModal } from '@rocket.chat/ui-client';
 
-import { e2e } from '../../../../app/e2e/client';
 import { settings } from '../../../../app/settings/client';
 import { fileUploadIsValidContentType } from '../../../../app/utils/client';
 import { getFileExtension } from '../../../../lib/utils/getFileExtension';
 import FileUploadModal from '../../../views/room/modals/FileUploadModal';
+import { e2e } from '../../e2ee';
 import { prependReplies } from '../../utils/prependReplies';
 import type { ChatAPI } from '../ChatAPI';
