+++ apps/meteor/client/lib/e2ee/helper.ts
