+++ apps/meteor/client/views/room/composer/messageBox/MessageBoxHint.spec.tsx
@@ -2,7 +2,7 @@ import { mockAppRoot } from '@rocket.chat/mock-providers';
 import { render, screen } from '@testing-library/react';
 
 import MessageBoxHint from './MessageBoxHint';
-import { E2ERoomState } from '../../../../../app/e2e/client/E2ERoomState';
+import { E2ERoomState } from '../../../../lib/e2ee/E2ERoomState';
 import { useRoom } from '../../contexts/RoomContext';
 import { useE2EERoomState } from '../../hooks/useE2EERoomState';
