+++ apps/meteor/client/views/room/composer/messageBox/MessageBoxHint.tsx
@@ -3,7 +3,7 @@ import type { ReactElement } from 'react';
 import { memo } from 'react';
 import { useTranslation, Trans } from 'react-i18next';
 
-import { E2ERoomState } from '../../../../../app/e2e/client/E2ERoomState';
+import { E2ERoomState } from '../../../../lib/e2ee/E2ERoomState';
 import { useRoom } from '../../contexts/RoomContext';
 import { useE2EERoomState } from '../../hooks/useE2EERoomState';
