+++ apps/meteor/client/views/room/contextualBar/RoomFiles/hooks/useFilesList.ts
@@ -2,9 +2,9 @@ import { Base64 } from '@rocket.chat/base64';
 import { useUserRoom, useEndpoint } from '@rocket.chat/ui-contexts';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 
-import { e2e } from '../../../../../../app/e2e/client/rocketchat.e2e';
 import { useScrollableRecordList } from '../../../../../hooks/lists/useScrollableRecordList';
 import { useComponentDidUpdate } from '../../../../../hooks/useComponentDidUpdate';
+import { e2e } from '../../../../../lib/e2ee/rocketchat.e2e';
 import type { FilesListOptions } from '../../../../../lib/lists/FilesList';
 import { FilesList } from '../../../../../lib/lists/FilesList';
 import { getConfig } from '../../../../../lib/utils/getConfig';
