+++ apps/meteor/client/views/room/hooks/useE2EERoomState.ts
@@ -1,7 +1,7 @@
 import { useMemo, useSyncExternalStore } from 'react';
 
 import { useE2EERoom } from './useE2EERoom';
-import type { E2ERoomState } from '../../../../app/e2e/client/E2ERoomState';
+import type { E2ERoomState } from '../../../lib/e2ee/E2ERoomState';
 
 export const useE2EERoomState = (rid: string) => {
 	const e2eRoom = useE2EERoom(rid);
