+++ apps/meteor/client/views/room/hooks/useE2EEState.ts
@@ -1,7 +1,7 @@
 import { useSyncExternalStore } from 'react';
 
-import { e2e } from '../../../../app/e2e/client';
-import type { E2EEState } from '../../../../app/e2e/client/E2EEState';
+import { e2e } from '../../../lib/e2ee';
+import type { E2EEState } from '../../../lib/e2ee/E2EEState';
 
 const subscribe = (callback: () => void): (() => void) => e2e.on('E2E_STATE_CHANGED', callback);
 const getSnapshot = (): E2EEState => e2e.getState();
