+++ apps/meteor/client/views/room/providers/hooks/useChatMessagesInstance.spec.ts
@@ -4,9 +4,9 @@ import { useUserId } from '@rocket.chat/ui-contexts';
 import { renderHook } from '@testing-library/react';
 
 import { useChatMessagesInstance } from './useChatMessagesInstance';
-import { E2ERoomState } from '../../../../../app/e2e/client/E2ERoomState';
 import { ChatMessages } from '../../../../../app/ui/client/lib/ChatMessages';
 import { useEmojiPicker } from '../../../../contexts/EmojiPickerContext';
+import { E2ERoomState } from '../../../../lib/e2ee/E2ERoomState';
 import { useUiKitActionManager } from '../../../../uikit/hooks/useUiKitActionManager';
 import { useRoomSubscription } from '../../contexts/RoomContext';
 import { useE2EERoomState } from '../../hooks/useE2EERoomState';
